requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==2.0.1
opencv-python==4.10.0.84
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pywin32==306
pyzmq==25.1.1
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
wcwidth==0.2.6
